+++ src/main.tsx
@@ -2,9 +2,9 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App.tsx";
 import { BrowserRouter } from "react-router-dom";
-import { GlobalStyle } from "./GlobalStyle.ts";
+import { GlobalStyle } from "./style/GlobalStyle.ts";
 import { ThemeProvider } from "styled-components";
-import { theme } from "./theme";
+import { theme } from "./style/theme.ts";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
